requirements.txt
click==8.0.1
Flask==2.0.1
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
PyWavelets==1.1.1
six==1.16.0
Werkzeug==2.0.1
WTForms==2.3.3
